+++ frontend/src/Components/HomePage/HomePage.js
@@ -1,14 +1,12 @@
 import React from "react";
 import CallToAction from "./Sections/CallToAction";
 import AboutUsSection from "./Sections/AboutUsSection";
-import AboutUsSection2 from "./Sections/AboutUsSection2";
 
 const HomePage = () => {
   return (
     <>
       <CallToAction />
-      {/* <AboutUsSection />0 */}
-      <AboutUsSection2 />
+      <AboutUsSection />
     </>
   );
 };
