--- frontend/src/Components/HomePage/Sections/AboutUsSection2.css
@@ -1,88 +1,0 @@
-.about-us-section {
-  background-image: url("../../../utils/cover/about-us-portrait-bg.png");
-  background-size: 100%;
-  background-position: center;
-  background-repeat: no-repeat;
-
-  color: #fff;
-  width: 100%;
-  height: 100%;
-
-  margin: 0;
-  padding-top: 10%;
-  padding-bottom: 10%;
-
-  display: block;
-  justify-content: left;
-  align-items: left;
-}
-
-.about-us-text {
-  display: flex;
-  flex-direction: column;
-  max-width: 50%;
-  justify-content: left;
-  margin: 0;
-}
-
-.about-us-section .about-us-portrait {
-  display: none;
-}
-
-@media (max-width: 1024px) {
-  .about-us-section {
-    background-image: url("../../../utils/cover/about-us-portrait-bg.png");
-    background-size: 150%;
-    background-repeat: no-repeat;
-    background-position: right;
-
-    color: #fff;
-    width: 100%;
-    height: 100%;
-
-    margin: 0;
-    padding-top: 10%;
-    padding-bottom: 10%;
-
-    display: block;
-    justify-content: left;
-    align-items: left;
-  }
-}
-
-@media (max-width: 480px) {
-  .about-us-section {
-    background-image: none;
-    padding: 0;
-  }
-
-  .about-us-section .about-us-portrait {
-    display: block;
-    width: 80%;
-    height: auto;
-    margin-top: 2rem;
-    margin-bottom: 1rem;
-  }
-
-  .about-us-text {
-    display: flex;
-    flex-direction: column;
-    margin-left: auto;
-    margin-right: auto;
-    max-width: 75%;
-    justify-content: center;
-    align-items: center;
-  }
-
-  .about-us-text h2 {
-    text-align: center;
-    font-size: 50px;
-    color: #daa41a;
-  }
-
-  .about-us-text p {
-    font-size: 18px;
-    width: 100%;
-    color: black;
-  }
-}
